Hotfix kb/gamepad turn and model rotation
Keyboard and gamepad turn was resulting in a jerk movement, disabling injection on aimX while the model is turning fix this.

Using 30.f on smoothing makes a nice and tight rotation, but when tabbing the emulator and coming back the movement feels sluggish. My guess is something due to the tickrate of the injector changing somehow while alt+tabbing. Using 5.f makes the rotation too sharp at first, but tabbing makes it feels like before (nice and tight).